requirements.txt
 openai 
fastapi 
uvicorn 
paicorn
pandas
dotenv-cpu
pydantic
